[#30] Refactor: migrate from recoil to jotai
