+++ src/stores/atoms/loginStateAtom.ts
@@ -1,6 +1,13 @@
-import { atom } from 'recoil';
+import { atom } from 'jotai';
 
-export const loginStateAtom = atom({
-  key: 'loginState',
-  default: false,
-});
+export const loginStateAtom = atom(false);
+
+export const readWriteLoginState = atom(
+  (get) => {
+    return get(loginStateAtom);
+  },
+
+  (get, set, newLoginState: boolean) => {
+    return set(loginStateAtom, newLoginState);
+  },
+);
